Remove unnecessary sleeps from chewie tests
This is a unit test, not an integration test. There is more work to be
done as we're replicating a lot of the state machine tests here, but
this at least cuts a chunk of time out of the test suite.
